SpeedUp by skiping GetEntries()
TChain->GetEntries() will open each root file in the filelist, get the number
of entries count and then sum them up. This really slow the process if you are
processing file though xrootd and have a lot list, also a huge waste of CPU
time. This is to remove the usage of nEvtTotal, which is taking as the max of
int for now.